requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP for concept extraction
spacy>=3.7.0
scikit-learn>=1.3.0

# API clients
arxiv>=2.0.0
biopython>=1.83

# Web scraping (for Google Scholar fallback)
beautifulsoup4>=4.12.0

# Caching & rate limiting
requests-cache>=1.1.0
ratelimit>=2.2.1

# Data validation
validators>=0.22.0

# API server (for React integration)
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
